third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/infoOnProgressEvent-response-type-blob-expected.txt

.onprogress handler on XMLHttpRequest objects with response type "blob" works to support progressive download content length information

You should see type, bubble, cancelable, eventPhase, target and current target.

Type: progress
Bubble: false
Cancelable: false
EventPhase: 2
Target: [object XMLHttpRequest]
Current target: [object XMLHttpRequest]
